refactors 'flyout-button' tests (#57572)
