+++ x-pack/legacy/plugins/siem/cypress/integration/smoke_tests/timeline/flyout_button.spec.ts
@@ -4,44 +4,34 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
+import { HOSTS_PAGE } from '../../../urls/navigation';
+import { waitForAllHostsToBeLoaded, dragFirstHostToTimeline } from '../../../tasks/hosts/all_hosts';
+import { loginAndWaitForPage } from '../../../tasks/login';
+import { openTimelineIfClosed, openTimeline } from '../../../tasks/siem_main';
 import {
   TIMELINE_FLYOUT_BODY,
   TIMELINE_NOT_READY_TO_DROP_BUTTON,
-} from '../../lib/timeline/selectors';
-import { ALL_HOSTS_WIDGET_DRAGGABLE_HOSTS } from '../../lib/hosts/selectors';
-import { HOSTS_PAGE } from '../../lib/urls';
-import { waitForAllHostsWidget } from '../../lib/hosts/helpers';
-import { loginAndWaitForPage } from '../../lib/util/helpers';
-import { drag } from '../../lib/drag_n_drop/helpers';
-import { createNewTimeline, toggleTimelineVisibility } from '../../lib/timeline/helpers';
+} from '../../../screens/timeline/main';
+import { createNewTimeline } from '../../../tasks/timeline/main';
 
 describe('timeline flyout button', () => {
   before(() => {
     loginAndWaitForPage(HOSTS_PAGE);
+    waitForAllHostsToBeLoaded();
   });
 
   afterEach(() => {
-    cy.get('[data-test-subj="kibanaChrome"]').then($page => {
-      if ($page.find('[data-test-subj="flyoutOverlay"]').length === 1) {
-        toggleTimelineVisibility();
-      }
-    });
-
+    openTimelineIfClosed();
     createNewTimeline();
   });
 
   it('toggles open the timeline', () => {
-    toggleTimelineVisibility();
-
+    openTimeline();
     cy.get(TIMELINE_FLYOUT_BODY).should('have.css', 'visibility', 'visible');
   });
 
   it('sets the flyout button background to euiColorSuccess with a 10% alpha channel when the user starts dragging a host, but is not hovering over the flyout button', () => {
-    waitForAllHostsWidget();
-
-    cy.get(ALL_HOSTS_WIDGET_DRAGGABLE_HOSTS)
-      .first()
-      .then(host => drag(host));
+    dragFirstHostToTimeline();
 
     cy.get(TIMELINE_NOT_READY_TO_DROP_BUTTON).should(
       'have.css',
