+++ x-pack/legacy/plugins/siem/cypress/screens/siem_main.ts
@@ -1,0 +1,9 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+export const MAIN_PAGE = '[data-test-subj="kibanaChrome"]';
+
+export const TIMELINE_TOGGLE_BUTTON = '[data-test-subj="flyoutOverlay"]';
