+++ x-pack/legacy/plugins/siem/cypress/screens/timeline/main.ts
@@ -32,3 +32,8 @@ export const TIMELINE_DATA_PROVIDERS_EMPTY =
 
 export const TIMELINE_DROPPED_DATA_PROVIDERS =
   '[data-test-subj="dataProviders"] [data-test-subj="providerContainer"]';
+
+export const TIMELINE_FLYOUT_BODY = '[data-test-subj="eui-flyout-body"]';
+
+export const TIMELINE_NOT_READY_TO_DROP_BUTTON =
+  '[data-test-subj="flyout-button-not-ready-to-drop"]';
